+++ e2e/todos/new/select-list-test.spec.ts
@@ -1,9 +1,13 @@
 import { test, expect } from '../../fixtures/index';
 import { v4 as uuidv4 } from 'uuid';
+import { deleteLists } from '../../helpers/teardown';
 
 let list: List;
+
 test.describe('creating a new task and selecting a list', () => {
     test.beforeEach(async ({ listAPI, page }) => {
+        await deleteLists();
+
         const testId = uuidv4();
         list = await listAPI.new({
             name: `List Name ${testId}`,
@@ -12,22 +16,27 @@ test.describe('creating a new task and selecting a list', () => {
         await page.goto('/');
         await page.waitForLoadState('networkidle');
     });
+
     test('it should create a task with a list', async ({ page }) => {
         const testId = uuidv4();
         const todoName = `Todo ${testId}`;
         await page.keyboard.press('t');
 
-        await page.waitForTimeout(500);
+        await page.waitForLoadState('networkidle');
+
         const dialogTitle = await page.getByTestId('dialog-title').textContent();
-        expect(dialogTitle).toBe('New Task');
-        const newTodoInput = await page.getByTestId('new-todo-input').locator('input').first();
+        expect(dialogTitle).toContain('New Task');
+
+        const newTodoInput = await page.getByRole('textbox', { name: 'What needs to be done?' });
         await newTodoInput.fill(todoName);
 
-        await page.locator('.v-field.v-field--appended.v-field--center-affix.v-field--no-label.v-field--variant-filled.v-theme--light > .v-field__field > .v-field__input').click();
+        await page.getByTestId('list-select').click();
         await page.getByRole('option', { name: list.name, exact: true }).click();
 
         const responsePromise = page.waitForResponse('**/api/todo');
         await page.getByTestId('create-todo-button').click();
+
+        await page.pause();
         const response = await responsePromise;
         expect(response.status()).toBe(200);
         const body: Todo = await response.json();
