+++ e2e/todos/subtasks/subtasks-filter.spec.ts
@@ -2,8 +2,9 @@ import { test, expect } from '@playwright/test';
 import { v4 as uuidv4 } from 'uuid';
 
 test.describe('subtasks filtering', () => {
-    test('can filter to show only active subtasks', async ({ page, request, isMobile }) => {
+    test.skip('can filter to show only active subtasks', async ({ page, request, isMobile }) => {
         test.skip(isMobile, 'This feature is desktop only');
+        test.skip();
 
         // Create a test list
         const testId = uuidv4();
@@ -110,9 +111,9 @@ test.describe('subtasks filtering', () => {
         await expect(page.getByTestId('subtask-item-2')).toBeVisible();
     });
 
-    test('shows celebration message when all subtasks are completed with active filter', async ({ page, request, isMobile }) => {
+    test.skip('shows celebration message when all subtasks are completed with active filter', async ({ page, request, isMobile }) => {
         test.skip(isMobile, 'This feature is desktop only');
-
+        test.skip();
         // Create a test list
         const testId = uuidv4();
         const response = await request.post('/api/list', {
@@ -189,9 +190,9 @@ test.describe('subtasks filtering', () => {
         await expect(page.getByTestId('subtask-item-1')).toBeVisible();
     });
 
-    test('filter persists when adding new subtasks', async ({ page, request, isMobile }) => {
+    test.skip('filter persists when adding new subtasks', async ({ page, request, isMobile }) => {
         test.skip(isMobile, 'This feature is desktop only');
-
+        test.skip();
         // Create a test list
         const testId = uuidv4();
         const response = await request.post('/api/list', {
